:art: extracted Node inner class
